+++ src/lexic/scanner/identifier.rs
@@ -34,12 +34,15 @@ fn scan_impl(chars: &Vec<char>, start_pos: usize, current: String, is_datatype: 
             is_datatype,
         ),
         _ => {
+            // start_pos is the position where the token ENDS, not where it STARTS,
+            // so this is used to retrieve the original START position of the token
+            let current_len = current.len();
             if let Some(token_type) = str_is_keyword(&current) {
-                LexResult::Some(token::new(current, start_pos as i32, token_type), start_pos)
+                LexResult::Some(token::new(current, start_pos - current_len, token_type), start_pos)
             } else if is_datatype {
-                LexResult::Some(token::new_datatype(current, start_pos as i32), start_pos)
+                LexResult::Some(token::new_datatype(current, start_pos - current_len), start_pos)
             } else {
-                LexResult::Some(token::new_identifier(current, start_pos as i32), start_pos)
+                LexResult::Some(token::new_identifier(current, start_pos - current_len), start_pos)
             }
         }
     }