+++ src/lexic/scanner/mod.rs
@@ -33,7 +33,7 @@ pub fn grouping_sign(c: char, _: &Vec<char>, start_pos: usize) -> Option<LexResu
         _ => return None,
     };
 
-    let token = token::new(c.to_string(), start_pos as i32, token_type);
+    let token = token::new(c.to_string(), start_pos, token_type);
     Some(LexResult::Some(token, start_pos + 1))
 }